feat: refactor image handling with Cloudinary integration and remove unused image component
