+++ .changeset/afraid-eels-send.md
@@ -1,0 +1,5 @@
+---
+"@mogeko/blog": minor
+---
+
+Refactor image handling with Cloudinary integration.
