+++ apps/blog/components/render.tsx
@@ -7,13 +7,13 @@ import { Suspense } from "react";
 import { twMerge } from "tailwind-merge";
 
 import { Icon } from "@/components/icon";
+import { Image } from "@/components/image";
 import { TableBox } from "@/components/table-box";
 import { RichText, plainText } from "@/components/text";
 import { Details, Summary } from "@/components/ui/accordion";
 import { ActionLink } from "@/components/ui/action-link";
 import { Card } from "@/components/ui/card";
 import { Heading } from "@/components/ui/heading";
-import { Image } from "@/components/ui/image";
 import { ListItem } from "@/components/ui/list";
 import { Loading } from "@/components/ui/loading";
 import { Separator } from "@/components/ui/separator";
