+++ apps/blog/next.config.ts
@@ -5,6 +5,7 @@ const nextConfig: NextConfig = {
     remotePatterns: [
       { protocol: "https", hostname: "**.notionusercontent.com" },
       { protocol: "https", hostname: "**.amazonaws.com" },
+      { protocol: "https", hostname: "res.cloudinary.com/mogeko/image/upload" },
     ],
   },
   logging: {
