+++ apps/blog/package.json
@@ -16,6 +16,7 @@
     "@shikijs/colorized-brackets": "^3.2.2",
     "@shikijs/transformers": "^3.2.2",
     "class-variance-authority": "^0.7.1",
+    "cloudinary": "^2.6.0",
     "clsx": "^2.1.1",
     "date-fns": "^4.1.0",
     "katex": "^0.16.22",
@@ -24,7 +25,6 @@
     "react-dom": "^19.1.0",
     "satori": "^0.12.2",
     "server-only": "^0.0.1",
-    "sharp": "^0.34.1",
     "shiki": "^3.2.2",
     "tailwind-merge": "^3.2.0"
   },
